Fix JooqProviderTest inheritance issue by using JUnit assume conditionals
